+++ milkman-graphql/src/main/java/milkman/plugin/graphql/domain/GraphqlAspect.java
@@ -1,0 +1,10 @@
+package milkman.plugin.graphql.domain;
+
+import lombok.Data;
+import milkman.domain.RequestAspect;
+
+@Data
+public class GraphqlAspect extends RequestAspect {
+	private String query = "";
+	private String variables = "{}";
+}